+++ src/pydiverse/pipedag/backend/table/sql.py
@@ -401,7 +401,7 @@ class SQLTableStore(BaseTableStore):
         early_not_null_possible: bool = False,
     ):
         if not early_not_null_possible:
-            for retry_iteration in range(3):
+            for retry_iteration in range(4):
                 # retry operation since it might have been terminated as a
                 # deadlock victim
                 try:
@@ -411,6 +411,8 @@ class SQLTableStore(BaseTableStore):
                         )
                     )
                 except sa.exc.SQLAlchemyError:
+                    if retry_iteration == 3:
+                        raise
                     time.sleep(retry_iteration * retry_iteration * 1.1)
         self.execute(AddPrimaryKey(table_name, schema, key_columns, name))
 
@@ -1429,12 +1431,20 @@ class SQLAlchemyTableHook(TableHook[SQLTableStore]):
         table_name = table.name
         schema = store.get_schema(stage_name).get()
         table_name, schema = store.resolve_aliases(table_name, schema)
-        return sa.Table(
-            table_name,
-            sa.MetaData(bind=store.engine),
-            schema=schema,
-            autoload_with=store.engine,
-        )
+        for retry_iteration in range(4):
+            # retry operation since it might have been terminated as a deadlock victim
+            try:
+                tbl = sa.Table(
+                    table_name,
+                    sa.MetaData(),
+                    schema=schema,
+                    autoload_with=store.engine,
+                )
+            except sa.exc.SQLAlchemyError:
+                if retry_iteration == 3:
+                    raise
+                time.sleep(retry_iteration * retry_iteration * 1.2)
+        return tbl
 
     @classmethod
     def lazy_query_str(cls, store, obj) -> str:
@@ -1471,11 +1481,13 @@ def _resolve_alias_ibm_db_sa(conn, table_name: str, schema: str, *, _iteration=0
             (tbl.c.creator == schema) & (tbl.c.name == table_name) & (tbl.c.TYPE == "A")
         )
     )
-    for retry_iteration in range(3):
+    for retry_iteration in range(4):
         # retry operation since it might have been terminated as a deadlock victim
         try:
             row = conn.execute(query).mappings().one_or_none()
         except sa.exc.SQLAlchemyError:
+            if retry_iteration == 3:
+                raise
             time.sleep(retry_iteration * retry_iteration)
     if row is not None:
         assert _iteration < 3, f"Unexpected recursion looking up {schema}.{table_name}"